register, otp, resetpassword
# Conflicts:
#	mvnw
#	src/main/java/com/kshrd/lumnov/controller/AuthController.java
#	src/main/java/com/kshrd/lumnov/exception/GlobalException.java
#	src/main/java/com/kshrd/lumnov/model/dto/request/AppUserRequest.java
#	src/main/java/com/kshrd/lumnov/model/dto/response/AppUserResponse.java
#	src/main/java/com/kshrd/lumnov/model/entity/AppUser.java
#	src/main/java/com/kshrd/lumnov/repository/AppUserRepository.java
#	src/main/java/com/kshrd/lumnov/service/AppUserService.java
#	src/main/java/com/kshrd/lumnov/service/impl/AppUserServiceImpl.java
#	src/main/resources/application.properties
#	src/main/resources/schema.sql

+++ src/main/java/com/kshrd/lumnov/model/entity/AppUser.java
@@ -18,7 +18,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AppUser implements UserDetails {
-  private Long appUserId;
+  private Integer appUserId;
   private String fullName;
   private Gender gender;
   private LocalDate dateOfBirth;
